+++ src/rs_embed/embedders/onthefly_remoteclip.py
@@ -26,13 +26,6 @@ from ..providers import ProviderBase
 # -----------------------------
 # Provider: Fetch S2 RGB
 # -----------------------------
-
-def _s2_rgb_u8_from_chw(s2_chw):
-    x = np.clip(np.asarray(s2_chw, dtype=np.float32), 0.0, 1.0)
-    return (x.transpose(1, 2, 0) * 255.0).astype(np.uint8)
-
-from .base import EmbedderBase
-from .meta import build_meta, temporal_to_range
 from ..providers.fetch import (
     fetch_s2_rgb_chw as _fetch_s2_rgb_chw_shared,
 )
@@ -42,6 +35,13 @@ from ..providers.resolution import (
 from ..tools.runtime import (
     resolve_device_auto_torch,
 )
+from .base import EmbedderBase
+from .meta import build_meta, temporal_to_range
+
+
+def _s2_rgb_u8_from_chw(s2_chw):
+    x = np.clip(np.asarray(s2_chw, dtype=np.float32), 0.0, 1.0)
+    return (x.transpose(1, 2, 0) * 255.0).astype(np.uint8)
 
 
 def _fetch_s2_rgb_chw(
