+++ src/rs_embed/embedders/onthefly_thor.py
@@ -23,6 +23,28 @@ from ..core.specs import (
 )
 from ..core.types import FetchResult
 from ..providers import ProviderBase
+from ..providers.fetch import (
+    fetch_collection_patch_chw as _fetch_collection_patch_chw,
+)
+from ..providers.fetch import (
+    fetch_s1_vvvh_raw_chw_with_meta as _fetch_s1_vvvh_raw_chw_with_meta,
+)
+from ..providers.fetch import (
+    normalize_s1_vvvh_chw as _normalize_s1_vvvh_chw,
+)
+from ..providers.resolution import (
+    is_provider_backend,
+)
+from ..tools.runtime import (
+    load_cached_with_device as _load_cached_with_device,
+)
+from ..tools.runtime import (
+    resolve_device_auto_torch as _resolve_device,
+)
+from .base import EmbedderBase
+from .config import model_config_value
+from .meta import build_meta, temporal_to_range
+
 
 def ensure_torch() -> None:
     try:
@@ -33,8 +55,6 @@ def ensure_torch() -> None:
 
 def pool_from_tokens(tokens, pooling):
     n = len(tokens)
-    h = int(n ** 0.5 - 1) if n > 1 else 0
-    # check h+1 squared equals n for CLS detection
     h2 = int((n - 1) ** 0.5)
     has_cls = n > 1 and h2 * h2 == n - 1
     patch = tokens[1:] if has_cls else tokens
@@ -53,26 +73,11 @@ def tokens_to_grid_dhw(tokens):
     has_cls = n > 1 and h2 * h2 == n - 1
     patch = tokens[1:] if has_cls else tokens
     p, d = patch.shape
-    hw = int(p ** 0.5)
+    hw = int(p**0.5)
     if hw * hw != p:
         raise ModelError(f"Patch token count {p} is not a perfect square.")
     return patch.reshape(hw, hw, d).transpose(2, 0, 1).astype("float32"), (hw, hw), has_cls
 
-from .base import EmbedderBase
-from .config import model_config_value
-from .meta import build_meta, temporal_to_range
-from ..providers.fetch import (
-    fetch_collection_patch_chw as _fetch_collection_patch_chw,
-    fetch_s1_vvvh_raw_chw_with_meta as _fetch_s1_vvvh_raw_chw_with_meta,
-    normalize_s1_vvvh_chw as _normalize_s1_vvvh_chw,
-)
-from ..providers.resolution import (
-    is_provider_backend,
-)
-from ..tools.runtime import (
-    load_cached_with_device as _load_cached_with_device,
-    resolve_device_auto_torch as _resolve_device,
-)
 
 _S2_SR_10_BANDS = ["B2", "B3", "B4", "B5", "B6", "B7", "B8", "B8A", "B11", "B12"]
 _S1_VVVH_BANDS = ["VV", "VH"]
